requirements.txt
streamlit
plotly
numpy
requests
python-dotenv
scikit-learn

# Testing
pytest
pytest-mock
